Profiled and memory checked the code. It is leak free (apart from bison, which seems a constant size leak), and some main smart-pointer overheads during parsing have been removed.

+++ solveridsmt/aggs/AggTypes.h
@@ -18,7 +18,7 @@ using namespace boost;
 
 #include "BigInteger.hh"
 #include "BigIntegerUtils.hh"
-typedef int Weight;
+typedef BigInteger Weight;
 
 namespace Aggrs{
 class WLV;
